refactor: split prompt files
